[WIP 2.4] migration of console-api queries to search-api
