+++ backend/test/csrf-attack-sample.html
@@ -13,11 +13,5 @@
       <input type="submit" />
     </form>
 
-    Test CSRF attack on: http://localhost:4000/search/console-api/graphql
-    <form action="http://localhost:4000/search/console-api/graphql" method="POST" enctype='text/plain'>
-      <input type="hidden" name='{"operationName": "saveSearch", "query": "mutation saveSearch($resource: JSON!) {\n savedSearch(resource: $resource)\n}\n", "variables":  {"resource": {"id": "161102530000", "name": "csrf attack", "description": "Created by csrf attack.", "searchText": "kind:pod"}}, "unused": "' value='"}' />
-      <input type="submit" />
-    </form>
-
   </body>
 </html>
