--- frontend/src/console-sdk/console-sdk.yml
@@ -1,13 +1,0 @@
-# Copyright (c) 2021 Red Hat, Inc.
-# Copyright Contributors to the Open Cluster Management project
-overwrite: true
-schema: http://localhost:4000/hcmuiapi/graphql
-documents: ./src/console-sdk/*.graphql
-generates:
-    src/console-sdk/console-sdk.ts:
-        plugins:
-            - typescript
-            - typescript-operations
-            - typescript-react-apollo
-        config:
-            skipTypename: true
