+++ frontend/src/routes/SearchPage/components/Modals/DeleteResourceModal.test.tsx
@@ -6,7 +6,7 @@ import userEvent from '@testing-library/user-event'
 import { MockedProvider } from '@apollo/client/testing'
 import { wait } from '../../../../lib/test-helper'
 import { DeleteResourceModal } from './DeleteResourceModal'
-import { DeleteResourceDocument, UserAccessDocument } from '../../../../console-sdk/console-sdk'
+import { DeleteResourceDocument, UserAccessDocument } from '../../../../search-sdk/search-sdk'
 import { SearchResultItemsDocument } from '../../../../search-sdk/search-sdk'
 
 describe('DeleteResourceModal', () => {
